Move benchmark module out of unit tests area. Move test helper functions into helpers.py instead of having them spread out in test modules that do not contain tests. Remove CLI from test modules because we use pytest now.
